feat: add fleet model tier policy
