+++ src/fleet/merge.rs
@@ -48,6 +48,7 @@ pub fn merge_instance_into_existing(
         ("repo", &config.repo),
         ("github_login", &config.github_login),
         ("model", &config.model),
+        ("model_tier", &config.model_tier),
         ("ready_pattern", &config.ready_pattern),
         ("command", &config.command),
         ("topic_binding_mode", &config.topic_binding_mode),
@@ -156,6 +157,7 @@ pub(super) fn build_instance_mapping(config: &InstanceYamlEntry) -> serde_yaml_n
         ("repo", &config.repo),
         ("github_login", &config.github_login),
         ("model", &config.model),
+        ("model_tier", &config.model_tier),
         ("ready_pattern", &config.ready_pattern),
         ("command", &config.command),
         ("skills_path", &config.skills_path),
