+++ src/mcp/tools.rs
@@ -96,7 +96,8 @@ pub(crate) fn def_create_instance() -> Value {
         "name": {"type": "string", "description": "Instance name (single instance) or base name (ignored when team is set — team name is used as prefix)"},
         "backend": {"type": "string", "description": "Backend CLI name: claude, agy, kiro-cli, codex, opencode"},
         "args": {"type": "string", "description": "Extra CLI arguments"},
-        "model": {"type": "string", "description": "Model override (e.g. --model flag)"},
+        "model": {"type": "string", "description": "Concrete model override (e.g. --model flag). Wins over model_tier."},
+        "model_tier": {"type": "string", "description": "Symbolic model tier key from fleet.yaml model_tiers (e.g. cheap/strong). Used when model is omitted; lets leads spawn mechanical-task workers on cheaper models (#2477)."},
         "working_directory": {"type": "string"},
         "branch": {"type": "string", "description": "Git branch — creates worktree if specified"},
         "task": {"type": "string", "description": "Initial task to inject after spawn"},
@@ -1002,6 +1003,7 @@ mod tests {
             ("create_instance", "backend", "instance_state/spawn.rs spawn backend"),
             ("create_instance", "args", "instance_state/spawn.rs extra cmd args"),
             ("create_instance", "model", "instance_state/spawn.rs --model flag"),
+            ("create_instance", "model_tier", "instance_state/spawn.rs fleet.yaml model_tier + api/handlers/instance.rs tier resolution"),
             ("create_instance", "working_directory", "instance_state/spawn.rs validated wd"),
             ("create_instance", "branch", "instance_state/spawn.rs worktree::create"),
             ("create_instance", "task", "instance_state/spawn.rs delayed inject"),
